backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
psycopg2-binary>=2.9
gunicorn>=22.0
